Add new module sources and package.json files
Only two modifications to the source files:
1. Explicitly import the modules named in the module's meta file
2. import a new 'y' module to get the shared Y global

Excplicit rollup modules were created that import all the rolled up modules.

Rollup modules that included logic in a source file (yui-base, yui, and loader)
were broken out into {module-name}-core.js and that new *-core module included
in the imports for the rollup.

A new module 'y' is added that exports a single would-be YUI instance with stubs for methods that _might_ be used, and full implementations for those that are likely to be or definitely used. This exported object takes the place of the YUI instance passed to the module body in YUI.add('foo', function (Y <-- HERE