Add event sub methods and version
event-base relies on Y.Env.add and .remove, which are wrappers for
el.addEventListener(...) and el.removeEventListener(...). I could have inlined
them in event-base, but easier to implement a simple pass through in the y
module.

Added a fixed version of '3.smugmug' to the y module's exported object so
Y.version will be populated for loader's internal references. This even though
it would be ideal if loader weren't even necessary. Maybe so.