Replace YUI with Y
Replace all references to the YUI global instance with the one Y imported from the shared 'y' module. In cases where code needed to check if a var was a YUI instance, change the comparison to `if ({thing} === Y)`, since there will only be one Y from the imported y module.